+++ R/check.R
@@ -1,10 +1,8 @@
 check_no_missing_values <- function(object) {
   missing <- vapply(object, anyNA, TRUE)
   
-  if(any(missing)) {
-    err(co(names(missing[missing]), 
-           "the following derived parameter includes missing values: %c",
-           "the following %n derived parameter%s include missing values: %c"))
-  }
-  object
+  if(any(missing))
+    err("the following derived parameters include missing values: ", 
+        cc(names(missing[missing]), " and "))
+  TRUE
 }