+++ NEWS.md
@@ -1,3 +1,4 @@
+- Replaced `checkr` and `err` dependencies for `chk`
 - Added mcmc_derive() for nlist, nlists and mcmc objects
 - Internal
   - Replace parameters() with pars()